Base Desktop Input working on windows
Lacking the function to set the keys, currently also using way to much space in key array,...

and Input and Display are merged in core because desktop input is managed through glfw and thus requiring access to some structures.

I don't want to implement windows/linux input base but may be needed, gainput seems like a pain to implement too.
